Implementata Servlet per piazzamento dell'ordine
modificati leggermente alcuni metodi e/o classi per adattarli al nuovo
contesto

+++ NibInk/src/main/java/com/model/ShippingMethod.java
@@ -4,7 +4,7 @@ public class ShippingMethod {
 	String name;
 	float amount;
 	String courier;
-	float percentage = 2.5;
+	float percentage = (float) 2.5;
 	
 	public String getName(){
 		return this.name;
